Add error checking for google_client_id and google_client_secret
Added checks that these settings are defined in trac.ini in the [openid]
section. Also refactored code to add a function to generate the login page
and repeatedly call this function rather than duplicating code over and over.

Note that the css_class setting is NOT defined in openid.css, though it ought
to be.