Fix bug requiring Google+ profile in order to login
A user name is only associated with a Google+ profile, so people
who don't have a profile received an error when trying to log in. This
just requires a guard and assigning the name as the email address in this
case.

Also cleaned up redirecting back to a valid login page when a user cancels
an authentication request at Google.com.